Move Currency tests into its own file
